requirements.txt
# ============================================================
# Core Deep Learning Frameworks
# ============================================================
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ============================================================
# YOLOv8 (Object Detection)
# ============================================================
ultralytics>=8.0.0

# ============================================================
# Computer Vision & Image Processing
# ============================================================
opencv-python>=4.7.0
albumentations>=1.3.0
Pillow>=9.5.0

# ============================================================
# Data Science & Numerical Computing
# ============================================================
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# ============================================================
# Visualization & Monitoring
# ============================================================
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.12.0

# ============================================================
# Progress Tracking & Utilities
# ============================================================
tqdm>=4.65.0
pyyaml>=6.0

# ============================================================
# Development & Testing (Optional)
# ============================================================
pytest>=7.3.0
ipykernel>=6.22.0
jupyter>=1.0.0

# ============================================================
# GPU Acceleration (Optional)
# ============================================================
# Install CUDA-specific PyTorch version from:
# https://pytorch.org/get-started/locally/
# 
# Example for CUDA 11.8:
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
